pyproject.toml
[project]
name = "twitch-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Marie Muravyeva",email = "marie.muraveva@gmail.com"}
]
requires-python = ">=3.11,<3.15"
dependencies = [
    "requests (>=2.33.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pandas (>=3.0.2,<4.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "apache-airflow (>=3.2.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
[dependency-groups]
dev = [
    "ipykernel (>=7.2.0,<8.0.0)",
    "apache-airflow (==3.2.0)"
]
